[MLIR] Fix 64-bit product during aten.view lowering
std::accumulate needs 64-bit init value to perform 64-bit arithmetic
on a list of integers.

+++ lib/Conversion/TorchToLinalg/DataMovement.cpp
@@ -38,7 +38,8 @@ using namespace mlir::torch;
 using namespace mlir::torch::Torch;
 
 static int64_t productReduce(ArrayRef<int64_t> a) {
-  return accumulate(a.begin(), a.end(), /*init=*/1, std::multiplies<int64_t>());
+  return accumulate(a.begin(), a.end(), /*init=*/static_cast<int64_t>(1),
+                    std::multiplies<int64_t>());
 }
 
 template <typename OpTy, typename OpAdaptor>
